rag_pipeline/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.20.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.0
lxml>=5.2.0
langchain>=0.2.11
langchain-huggingface>=0.0.3
huggingface_hub>=0.14.1
sentence-transformers>=2.2.2
transformers>=4.41.0
torch>=2.1.0
supabase>=2.5.0
psycopg2-binary>=2.9.6
numpy>=1.26.0
neo4j>=5.15.0


